+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project does not adhere to [Semantic Versioning](http://semver.org/spec/v2.
 ### Changed
 - Improve `--help` output: Hide some redundant options, add a bug reporting URL
   and explain the purpose of `mdcat`.
+- Reduce dependencies and thus build time
 
 ## [0.5.0] – 2018-01-27
 ### Added